requirements.txt
Django>=4.2.0
huggingface_hub>=0.17.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
numpy>=1.21.0
pandas>=1.3.0